+++ include/trw/syntax.h
@@ -29,6 +29,18 @@ namespace TemplateRenderWizard
         public:
             Tree();
             void initializeDefaults() override;
+
+        protected:
+            SyntaxTree::Syntax::SyntaxElement* beforeParse(std::list<SyntaxTree::Syntax::SyntaxElement*>*) override;
+
+        private:
+            std::list<SyntaxTree::Syntax::SyntaxElement*>* filterBrackets(std::list<SyntaxTree::Syntax::SyntaxElement*>*);
+            std::list<SyntaxTree::Syntax::SyntaxElement*>* filterMathHigh(std::list<SyntaxTree::Syntax::SyntaxElement*>*);
+            std::list<SyntaxTree::Syntax::SyntaxElement*>* filterMathLow(std::list<SyntaxTree::Syntax::SyntaxElement*>*);
+
+            SyntaxTree::Syntax::SyntaxElement* parseExpr(std::list<SyntaxTree::Syntax::SyntaxElement*>*);
+
+            bool isExprToken(SyntaxTree::Syntax::SyntaxElement*);
         };
     }
 }