+++ src/trw/tokens/double_dot.cpp
@@ -1,0 +1,14 @@
+#include <trw.h>
+#include <syntax-tree-lib.h>
+
+namespace TemplateRenderWizard::Token
+{
+    DoubleDot::DoubleDot(int line, int column) : SyntaxTree::Token::Token(line, column, nullptr) {
+        this->line = line;
+        this->column = column;
+    }
+
+    int DoubleDot::getType() {
+        return Type::DoubleDotType;
+    }
+}