+++ src/trw/tokens/file_path.cpp
@@ -1,0 +1,13 @@
+#include <trw.h>
+#include <io-buffer.h>
+#include <syntax-tree-lib.h>
+
+namespace TemplateRenderWizard::Token
+{
+    FilePath::FilePath(int line, int column, IOBuffer::IOReader *reader) : SyntaxTree::Token::Token(line, column, reader) {
+    }
+
+    int FilePath::getType() {
+        return Type::FilePathType;
+    }
+}